Move to contrib dir; add explicit spidev and RPi.GPIO deps
