+++ setup.py
@@ -110,6 +110,8 @@ __requires__ = [
     ##############
     "monotonic>=1.1,<1.2",
     "pyinotify>=0.9.6,<0.10",
+    "RPi.GPIO>=0.7.0<0.8.0",
+    "spidev>=3.5<3.6",
 
     #########
     # Pulse #
